backend/question-service/server/app/requirements.txt
annotated-types==0.5.0
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
docker==6.1.3
exceptiongroup==1.1.3
fastapi==0.103.0
h11==0.14.0
idna==3.4
lxml==4.9.3
packaging==23.1
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
websocket-client==1.6.2
